refactor the event logic out for android/ipad
It's important that if we're on keyup and we triggered on a comma that the comma is removed. The other characters (spaces/newline) are dealt with by jquery's trim.

Also refactored a bit of the logic to do with whether we should tag, as it was a bit of a mess. Think there's some work to do to correct the logic there, but it's 6 years old so not sure.

If you hit enter with text in the input, then we won't submit the form anymore